+++ NEWS.md
@@ -18,9 +18,13 @@ author and date.
 * Compatible with ggplot2 charts and pngs saved to disk.
 * "" alt text warning for decorative images only.
 
-#### handle_rmd_path
+#### handle_rmd_path (Not exported)
 
-* Not exported. Handler checks rmd path is valid.
+* Handler checks rmd path is valid.
+
+#### find_all_imgs (Not exported)
+
+* Helper func to find all images within an Rmd.
 
 #### sus_alt